requirements.txt
tensorflow
scipy
pillow
flask